+++ packages/plugin-headings/web/src/types.js
@@ -1,0 +1,1 @@
+export const HEADINGS_DROPDOWN_TYPE = 'HeadingsDropdown';